+++ HISTORY.rst
@@ -102,6 +102,8 @@
   Avoids keeping long lived objects in memory.
 - Consolidate app module loading using AppsModuleLoaderMixin.
 - Remove usage of django.utils.six.
+- Add django-silk as a development dependency.
+  Add a development setting for django-silk.
 
 3.4.9 (2020-05-26)
 ==================